moved files around to follow the MVC architecture.
