+++ src/main/java/com/soen387/controller/HelloServlet.java
@@ -1,3 +1,5 @@
+package com.soen387.controller;
+
 import java.io.*;
 import javax.servlet.http.*;
 import javax.servlet.annotation.*;
