+++ src/main/java/com/soen387/dataaccess/DBConnection.java
@@ -1,4 +1,4 @@
-package com.soen387.dao;
+package com.soen387.dataaccess;
 
 import java.io.IOException;
 import java.io.InputStream;
